Fix: 모임 수정 로직 오류 해결 (#251)
* Fix: 모임 수정 로직 오류 해결

* Fix: build gradle 수정

* Fix: Test 오류 해결

* Fix: 모임 수정 테스트 추가


* Fix: 모임 상세 조회 멤버 변수명 변경


---------

+++ src/main/java/com/boardgo/common/utils/S3Service.java
@@ -1,5 +1,7 @@
 package com.boardgo.common.utils;
 
+import static com.boardgo.common.constant.S3BucketConstant.*;
+
 import com.amazonaws.SdkClientException;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.ObjectMetadata;
@@ -52,6 +54,9 @@ public class S3Service {
 
     public void deleteFile(String fileName) {
         try {
+            if (fileName.startsWith(BOARDGAME)) {
+                return;
+            }
             amazonS3Client.deleteObject(bucket, fileName);
         } catch (SdkClientException sce) {
             throw new CustomS3Exception(sce.getMessage());
